requirements.txt
requests>=2.32.5,<3.0.0
pytz>=2025.2,<2026.0
PyYAML>=6.0.3,<7.0.0
fastmcp>=2.12.0,<2.14.0
websockets>=13.0,<14.0
feedparser>=6.0.10,<7.0.0
beautifulsoup4>=4.12.0,<5.0.0
python-dateutil>=2.8.0,<3.0.0
